+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 ## Unreleased
 
+### Changed
+- Fixed a bug where querying for products sometimes would be broken.
+
 ### Fixed
 - Fixed a bug where only fields in the first layout tab were visible when editing a product. ([#8](https://github.com/craftcms/commerce-digital-products/issues/8))