+++ CHANGELOG.md
@@ -4,9 +4,10 @@ Changelog
 ## Unreleased
 
 ### Changed
-- Fixed a bug where querying for products sometimes would be broken.
+- Changed the label for the digital products field to „Digital Products”. ([#13](https://github.com/craftcms/commerce-digital-products/issues/13))
 
 ### Fixed
+- Fixed a bug where querying for products sometimes would be broken.
 - Fixed a bug where only fields in the first layout tab were visible when editing a product. ([#8](https://github.com/craftcms/commerce-digital-products/issues/8))
 - Fixed a bug where product element slugs would not work. ([#9](https://github.com/craftcms/commerce-digital-products/issues/9))
 - Fixed a bug where querying for licenses would not work in some cases. ([#11](https://github.com/craftcms/commerce-digital-products/issues/11))